Add TextureFrames.cpp & fix frame size cast
Add Renderer\TextureFrames.cpp to the project and include ..\Build\Assets\Animations\bat_idle.json as a None asset. Reorganize project filters/source listings (moved headers into categories and removed a redundant Renderer.h include). Fix TextureFrames.cpp to cast column/row counts to float when computing frame size to avoid integer division.

+++ Engine/Renderer/TextureFrames.cpp
@@ -35,7 +35,7 @@ namespace nu{
 
 		Vector2 textureSize = m_texture->GetSize();
 
-		m_frameSize = textureSize / Vector2{ m_numColumns, m_numRows };
+		m_frameSize = textureSize / Vector2{ static_cast<float>(m_numColumns), static_cast<float>(m_numRows) };
 
 		return true;
 	}
